used ajax for post creation

+++ controllers/post_controller.js
@@ -6,13 +6,28 @@ module.exports.create=async function(req,res){
 
  try {
 
-    await Post.create({
+    let post=await Post.create({
 
         content:req.body.content,
         user:req.user._id
 
 
     });
+
+    if(req.xhr){
+
+      post= await post.populate('user','name').execPopulate();
+
+      return res.status(200).json({
+
+         data:{
+
+            post:post
+         },
+
+         message:"Post Created"
+      });
+    }
 
     req.flash('success','Post Published!'); 
 
@@ -40,6 +55,19 @@ module.exports.destroy=async function(req,res){
     if(post.user==req.user.id){
 
         post.remove();
+
+        if(req.xhr){
+
+         return res.status(200).json({
+
+            data:{
+
+               post_id:req.params.id
+            },
+
+            message:"Post deleted successfully"
+         })
+        }
 
        await Comment.deleteMany({post:req.params.id});
